requirements.txt
click
pre-commit
rich
python-gitlab
